+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Plugin can now handle setting overrides which require suite reload. This
+  is an API change coming in Defensics 2021.06.
 
 ## v1.2.0 - 2021-03-22
 ### Changed